Adding the base animation with some blue prints
I have added 9 different animations set up in an Animation Source, the respective blueprints utilized in the video(This may be a source of merging error, and I believe 39 total pixel art animation sprites folders of all animations possible for a Legend of Zelda style Knight. I have also added one other character that hasnt been set up with any animation that is a red ninja. This is all located in Content> Animation. I also added the 4 different background images the player can play in of a Tokyo street.